requirements.txt
icalevents
numpy
vobject
lxml
schedule
Pillow
requests
devtools

